Optimize
Rewrite the code building functions. Then define a cost function (with weighted load cases) and write an optimization function able to solve for different number of elements by side of the cube. This work well with 1 element but failed at 2 elements. Try to resolve this to have some results on Monday